refactor(agent-runtime): extract tool traits into corvus-traits
