+++ clients/agent-runtime/crates/corvus-traits/Cargo.toml
@@ -13,3 +13,4 @@ async-trait = "0.1"
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0", default-features = false, features = ["std"] }
 tokio = { version = "1.42", default-features = false, features = ["sync", "macros", "rt"] }
+tracing = { version = "0.1", default-features = false }
