+++ clients/agent-runtime/crates/corvus-traits/src/lib.rs
@@ -2,3 +2,4 @@ pub mod channels;
 pub mod memory;
 pub mod security;
 pub mod testing;
+pub mod tools;
